server/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
cloudinary==1.41.0
colorama==0.4.6
Deprecated==1.2.15
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
limits==3.14.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
